requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
defusedxml==0.5.0
diff-match-patch==20121119
Django==1.11.3
django-cors-headers==2.1.0
django-crispy-forms==1.6.1
django-filter==1.0.4
django-formtools==2.0
django-guardian==1.4.9
django-redis==4.9.0
django-reversion==2.0.0
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
drf-extensions==0.3.1
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.9.2
idna==2.6
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
Markdown==2.6.8
MarkupSafe==1.0
mysqlclient==1.3.10
oauthlib==2.1.0
odfpy==1.3.6
olefile==0.45.1
openpyxl==2.5.3
Pillow==4.2.1
pycryptodome==3.6.3
PyJWT==1.6.4
python3-openid==3.1.0
pytz==2018.4
PyYAML==3.12
redis==2.10.6
requests==2.18.4
requests-oauthlib==1.0.0
six==1.10.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
xlrd==1.1.0
XlsxWriter==0.9.8
xlwt==1.2.0
